Feat/seongho/comment alram #107

+++ server/src/main/java/com/hobbyt/domain/search/dto/SearchRequest.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class SearchRequest {
 	private String keyword;
-	private int offset;
+	private long offset;
 	private int limit;
 	private OrderBy orderBy;
